Logica para crear cuentas cuando se crea el socio y adema la logica para crear cuotas.

+++ application/classes/Controller/Socios.php
@@ -8,7 +8,7 @@ class Controller_Socios extends Controller_Template_Base
 
   public function before(){
     parent::before();
-    // Podria verificar el ROl del usuario y mostrar una panta que 
+    // Podria verificar el ROl del usuario y mostrar una panta que
     // que informe que no está autorizado a ver este recurso
     // Fix manual para fechas:
       if(isset($_POST['fecha_nacimiento']))
@@ -39,7 +39,7 @@ class Controller_Socios extends Controller_Template_Base
     $persona = ORM::factory('Persona');
     $socio = ORM::factory('Socio');
     if (isset($_POST) && Valid::not_empty($_POST))
-    {      
+    {
       // Factory es un patron de diseño, tener en cuenta.
       $post = Validation::factory($_POST)
               ->rule('nombre','not_empty')
@@ -70,11 +70,11 @@ class Controller_Socios extends Controller_Template_Base
           catch (ORM_Validation_Exception $e)
           {
             $errors = $e->errors('socio');
-          }          
-        } 
+          }
+        }
         catch (ORM_Validation_Exception $e)
         {
-          $errors = $e->errors('persona');          
+          $errors = $e->errors('persona');
         }
       }
       else
@@ -155,13 +155,28 @@ class Controller_Socios extends Controller_Template_Base
   public function action_delete()
   {
     // Borramos el socio
-    $id = $this->request->param('id');
-    $socio = ORM::factory('Socio',$id);
-    $persona = $socio->persona;
-    # TODO agregar control de error al borrar
-    $socio->delete();
-    $persona->delete();
-    $this->redirect('socios/index');
+    try {
+      $id = $this->request->param('id');
+      if (!$id)
+      {
+        # Como estamos dentro de un try/catch, cuando generemos la nueva excepcion
+        # se tomará la rama del catch donde haremos el manejo del error y mostraremos el mensaje.
+        throw new Exception("No existe el registro para el id solicitado o no exite ningun id", 1);
+      }
+      $socio = ORM::factory('Socio',$id);
+      $persona = $socio->persona;
+      if ($persona->tiene_cuenta())
+      {
+        throw new Exception("Tiene Plan de Cuenta", 1);        
+      }
+      # TODO agregar control de error al borrar
+      $socio->delete();
+      $persona->delete();
+      $this->redirect('socios/index');
+      
+    } catch (Exception $e) {
+      $this->redirect('socios/index');      
+    }
   }
   public function action_consulta()
   {
